+++ Bencodex.Tests/Misc/ImplicitConversionTest.cs
@@ -41,8 +41,11 @@ namespace Bencodex.Tests.Misc
         {
             var binary = new Binary(new byte[] { 0x01, 0x02, 0x03 });
 
-            byte[] bytes = binary;
-            Assert.Equal(bytes, binary.Value);
+            ImmutableArray<byte> immutable = binary;
+            Assert.Equal(immutable, binary.ByteArray);
+
+            byte[] mutable = binary;
+            Assert.Equal(mutable, binary.ToByteArray());
         }
 
         [Fact]